pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "fastapi-app"
version = "1.0.0"
description = "A simple FastAPI application"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]

dependencies = [
    "langchain-core>=0.1.0",
    "langchain-mcp-adapters>=0.0.3",
    "langgraph>=0.0.26",
    "langchain-openai>=0.0.1",
    "langchain-google-genai>=0.0.1",
    "langchain-anthropic>=0.1.1",
    "langchain-groq>=0.0.1",
    "fastapi>=0.104.1",
    "uvicorn>=0.25.0",
    "httpx>=0.25.0",
    "aiohttp>=3.9.1",
    "python-dotenv>=1.0.0",
    "pydantic>=2.4.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
    "pytest-asyncio>=0.21.1",
]

[tool.setuptools]
packages = ["app"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.black]
line-length = 88
target-version = ["py39"]

[tool.isort]
profile = "black"
line_length = 88